refactored webpack stream to use pump
